# gen/http/openapi.json
{"swagger":"2.0","info":{"title":"WantGo API","description":"This is WantGoAPI","license":{"name":"MIT","url":"https://github.com/goadesign/goa/blob/master/LICENSE"},"version":""},"host":"127.0.0.1:8081","consumes":["application/json","application/xml","application/gob"],"produces":["application/json","application/xml","application/gob"],"paths":{"/card":{"post":{"tags":["WantGo"],"summary":"postCardInfo WantGo","operationId":"WantGo#postCardInfo","parameters":[{"name":"Authorization","in":"header","description":"Auth token","required":false,"type":"string","pattern":"^Bearer [^ ]+$"},{"name":"PostCardInfoRequestBody","in":"body","required":true,"schema":{"$ref":"#/definitions/WantGoPostCardInfoRequestBody","required":["cardAuthor","cardTitle","cardDescription"]}}],"responses":{"200":{"description":"OK response."}},"schemes":["http"]}},"/card-list":{"get":{"tags":["WantGo"],"summary":"getSimpleCardList WantGo","operationId":"WantGo#getSimpleCardList","parameters":[{"name":"Authorization","in":"header","required":false,"type":"string","pattern":"^Bearer [^ ]+$"}],"responses":{"200":{"description":"OK response.","schema":{"type":"array","items":{"$ref":"#/definitions/simpleCardResponse"}}}},"schemes":["http"]}},"/card/{cardId}":{"get":{"tags":["WantGo"],"summary":"getCardInfo WantGo","operationId":"WantGo#getCardInfo","parameters":[{"name":"cardId","in":"path","required":true,"type":"string"},{"name":"Authorization","in":"header","required":false,"type":"string","pattern":"^Bearer [^ ]+$"}],"responses":{"200":{"description":"OK response.","schema":{"$ref":"#/definitions/WantGoGetCardInfoResponseBody"}}},"schemes":["http"]},"put":{"tags":["WantGo"],"summary":"putCardInfo WantGo","operationId":"WantGo#putCardInfo","parameters":[{"name":"cardId","in":"path","required":true,"type":"string"},{"name":"Authorization","in":"header","required":false,"type":"string","pattern":"^Bearer [^ ]+$"},{"name":"PutCardInfoRequestBody","in":"body","required":true,"schema":{"$ref":"#/definitions/WantGoPutCardInfoRequestBody","required":["cardAuthor","cardTitle","cardDescription"]}}],"responses":{"200":{"description":"OK response."}},"schemes":["http"]},"delete":{"tags":["WantGo"],"summary":"deleteCardInfo WantGo","operationId":"WantGo#deleteCardInfo","parameters":[{"name":"cardId","in":"path","description":"card id","required":true,"type":"integer"},{"name":"Authorization","in":"header","required":false,"type":"string","pattern":"^Bearer [^ ]+$"}],"responses":{"200":{"description":"OK response."}},"schemes":["http"]}},"/openapi.json":{"get":{"tags":["WantGo"],"summary":"Download ./gen/http/openapi.json","operationId":"WantGo#/openapi.json","responses":{"200":{"description":"File downloaded","schema":{"type":"file"}}},"schemes":["http"]}}},"definitions":{"WantGoGetCardInfoResponseBody":{"title":"WantGoGetCardInfoResponseBody","type":"object","properties":{"cardAuthor":{"type":"string","default":"default","example":"Iusto qui dolores."},"cardDescription":{"type":"string","default":"default","example":"Illum et totam nam."},"cardTitle":{"type":"string","default":"default","example":"Dolore et sed quasi labore ratione laboriosam."},"imageUrl":{"type":"string","default":"default","example":"Rerum culpa aut corporis voluptatem deleniti."},"locationAddress":{"type":"string","default":"default","example":"Sit nobis aliquam non delectus vel."},"locationUrl":{"type":"string","default":"default","example":"Quia sit repudiandae non porro totam sunt."},"tags":{"type":"array","items":{"type":"string","example":"Quod tenetur sed."},"example":["Ab voluptas quia ut in.","Consequatur consequatur."]}},"example":{"cardAuthor":"Repellendus dolor qui fugiat quae.","cardDescription":"Ipsum quaerat enim est exercitationem vitae.","cardTitle":"Sed dignissimos est laudantium qui et.","imageUrl":"Repellat cumque eos fugiat quidem.","locationAddress":"Quam saepe inventore.","locationUrl":"Aperiam necessitatibus harum fuga velit placeat.","tags":["Minima voluptatem aut sit quam.","Nulla odit placeat et ut debitis.","Id non."]}},"WantGoPostCardInfoRequestBody":{"title":"WantGoPostCardInfoRequestBody","type":"object","properties":{"cardAuthor":{"type":"string","example":"Voluptatem sit provident asperiores."},"cardDescription":{"type":"string","example":"Ea dolorem eos assumenda molestiae libero eaque."},"cardTitle":{"type":"string","example":"Earum autem quam."},"imageUrl":{"type":"string","default":"default","example":"Inventore minima et."},"locationAddress":{"type":"string","default":"default","example":"Iusto saepe qui."},"locationUrl":{"type":"string","default":"default","example":"Explicabo vero ut ipsam."},"tags":{"type":"array","items":{"type":"string","example":"Quis labore."},"example":["Reiciendis ad aut.","Aut autem et sequi."]}},"example":{"cardAuthor":"Quam repellendus voluptates qui.","cardDescription":"Laborum nihil dolore voluptatem enim.","cardTitle":"Harum natus accusantium fugiat qui quidem rerum.","imageUrl":"Hic rerum esse.","locationAddress":"Quas nemo et.","locationUrl":"Sunt occaecati molestiae sunt.","tags":["Illo ad.","Facilis necessitatibus sed officiis."]},"required":["cardAuthor","cardTitle","cardDescription"]},"WantGoPutCardInfoRequestBody":{"title":"WantGoPutCardInfoRequestBody","type":"object","properties":{"cardAuthor":{"type":"string","example":"Quas et molestias sit vel laboriosam omnis."},"cardDescription":{"type":"string","example":"Dolore dolore necessitatibus necessitatibus et atque."},"cardTitle":{"type":"string","example":"Eos et qui."},"imageUrl":{"type":"string","default":"default","example":"Similique voluptatem eveniet."},"locationAddress":{"type":"string","default":"default","example":"Molestiae sit dolorum."},"locationUrl":{"type":"string","default":"default","example":"Quia impedit possimus tempora in qui."},"tags":{"type":"array","items":{"type":"string","example":"Magnam laborum aut quaerat."},"example":["Ipsam recusandae eaque ut eligendi et quae.","Temporibus rerum.","Soluta aliquid facere consequuntur.","Corporis quia dignissimos."]}},"example":{"cardAuthor":"Maxime ducimus molestias.","cardDescription":"Eos dolorum.","cardTitle":"Consequatur sapiente maiores minus eaque ullam.","imageUrl":"Autem blanditiis doloremque est ipsam.","locationAddress":"Dolor et tempore.","locationUrl":"Ipsum nemo aut optio cumque.","tags":["Libero non sit.","Enim voluptatem laudantium qui beatae.","Ipsa est quis magni."]},"required":["cardAuthor","cardTitle","cardDescription"]},"simpleCardResponse":{"title":"simpleCardResponse","type":"object","properties":{"cardAuthor":{"type":"string","default":"default","example":"Et placeat debitis consectetur voluptates nesciunt ut."},"cardId":{"type":"integer","default":0,"example":836421000872514727,"format":"int64"},"cardTitle":{"type":"string","default":"default","example":"Molestias voluptatum."},"imageUrl":{"type":"string","default":"default","example":"Qui quod."}},"example":{"cardAuthor":"Dolorem accusamus quam vitae magni.","cardId":4273863744033214172,"cardTitle":"Ipsum accusantium aut officiis ut voluptates.","imageUrl":"Est dicta aliquam delectus sapiente maiores culpa."}}}}